input-box-to-set-amount-to-pay
dah ada input box untuk set amount to pay
